+++ CHANGELOG.md
@@ -11,6 +11,12 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v3.?.? (2021-??-??)
+
+#### :bug: Bug Fix
+
+* Fixed a bug with removing a modifier from the passed node `super/i-block/modules/block`
+
 ## v3.8.2 (2021-10-26)
 
 #### :bug: Bug Fix